src/requirements.txt
apns2==0.7.1
asgiref==3.2.10
backcall==0.2.0
billiard==3.6.3.0
celery==5.0.5
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
cryptography==3.1.1
decorator==4.4.2
defusedxml==0.6.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.16
django-allauth==0.42.0
django-celery-beat==2.1.0
django-celery-results==2.0.1
django-ckeditor==6.0.0
django-cors-headers==3.5.0
django-environ==0.4.5
django-extensions==3.0.9
django-filter==2.4.0
django-geoposition==0.3.0
django-js-asset==1.2.2
django-redis==4.12.1
django-rest-auth==0.9.5
django-social-auth==0.7.28
django-taggit==1.3.0
django-timezone-field==4.1.1
djangorestframework==3.12.0
et-xmlfile==1.0.1
geographiclib==1.50
geopy==2.0.0
gunicorn==20.0.4
h2==2.6.2
hpack==3.0.0
http-ece==1.1.0
httplib2==0.18.1
hyper==0.7.0
hyperframe==3.2.0
idna==2.10
importlib-metadata==2.0.0
ipdb==0.13.4
ipython==7.18.1
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.17.2
joblib==0.17.0
oauth2==1.9.0.post1
oauthlib==3.1.0
openpyxl==3.0.5
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.7
psycopg2==2.8.6
ptyprocess==0.6.0
py-vapid==1.7.1
pycparser==2.20
Pygments==2.7.1
PyJWT==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.1
python-openid==2.2.5
python3-openid==3.2.0
pytz==2020.1
pywebpush==1.11.0
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
scikit-learn==1.0.2
six==1.15.0
sklearn==0.0
sqlparse==0.4.2
static3==0.7.0
threadpoolctl==2.1.0
traitlets==5.0.4
wcwidth==0.2.5
zipp==3.3.0
django-clickhouse==1.1.2
timezonefinder==5.1.1
channels==2.3.0
channels-redis==2.4.0
daphne==2.5.0
simplejson==3.17.5
django-silk==4.0.0
pipe==1.6.0
drf-yasg==1.20.0
PyYAML==6.0
python-telegram-bot==13.11
jsonschema==4.5.1
sentry-sdk==1.9.5
